k8s adapter uses helper to parse json data
