+++ lib/ood_core/job/adapters/kubernetes/helper.rb
@@ -1,0 +1,132 @@
+class OodCore::Job::Adapters::Kubernetes::Helper
+
+
+
+  def info_from_json(pod_json: nil, service_json: nil, secret_json: nil)
+    pod_hash = pod_info_from_json(pod_json)
+    service_hash = service_info_from_json(service_json)
+    secret_hash = secret_info_from_json(secret_json)
+
+    # can't just use deep_merge bc we don't depend *directly* on rails
+    pod_hash[:native] = pod_hash[:native].merge(service_hash[:native])
+    pod_hash[:native] = pod_hash[:native].merge(secret_hash[:native])
+    OodCore::Job::Info.new(pod_hash)
+  end
+
+
+
+  private 
+
+  def pod_info_from_json(json_data)
+    return {} if json_data.nil?
+
+    # passing json_data around like it's OK, probably should check for nil?
+    id = json_data.dig(:metadata, :name).to_s
+    {
+      id: id,
+      job_name: name_from_metadata(json_data.dig(:metadata)),
+      status: pod_status_from_json(json_data),
+      job_owner: json_data.dig(:metadata, :namespace).to_s,
+      submission_time: submission_time(json_data),
+      dispatch_time: dispatch_time(json_data),
+      wallclock_time: wallclock_time(json_data),
+      native: {
+        host: json_data.dig(:status, :hostIP)
+      }
+    }
+  end
+
+  def name_from_metadata(metadata)
+    name = metadata.dig(:labels, :'app.kubernetes.io/name')
+    name = metadata.dig(:labels, :'k8s-app') if name.nil?
+    name = metadata.dig(:name) if name.nil? # pod-id but better than nil?
+    name
+  end
+
+  def service_info_from_json(json_data)
+    # .spec.ports[0].nodePort
+    ports = json_data.dig(:spec, :ports)
+    {
+      native:
+        {
+          port: ports[0].dig(:nodePort)
+        }
+    }
+  rescue
+    empty_native
+  end
+
+  def secret_info_from_json(json_data)
+    raw = json_data.dig(:data, :password)
+    {
+      native:
+        {
+          password: Base64.decode64(raw)
+        }
+    }
+  rescue
+    empty_native
+  end
+
+  def empty_native
+    {
+      native: {}
+    }
+  end
+
+  def dispatch_time(json_data)
+    status = pod_status_from_json(json_data)
+    state_data = json_data.dig(:status, :containerStatuses)[0].dig(:state)
+    date_string = nil
+
+    if status == 'completed'
+      date_string = state_data.dig(:terminated, :startedAt)
+    elsif status == 'running'
+      date_string = state_data.dig(:running, :startedAt)
+    end
+
+    date_string.nil? ? nil : DateTime.parse(date_string).to_time.to_i
+  end
+
+  def wallclock_time(json_data)
+    status = pod_status_from_json(json_data)
+    state_data = json_data.dig(:status, :containerStatuses)[0].dig(:state)
+    start_time = dispatch_time(json_data)
+    return nil if start_time.nil?
+
+    et = end_time(status, state_data)
+
+    et.nil? ? nil : et - start_time
+  end
+
+  def end_time(status, state_data)
+    if status == 'completed'
+      end_time_string = state_data.dig(:terminated, :finishedAt)
+      et = DateTime.parse(end_time_string).to_time.to_i
+    elsif status == 'running'
+      et = DateTime.now.to_time.to_i
+    else
+      et = nil
+    end
+
+    et
+  end
+
+  def submission_time(json_data)
+    str = json_data.dig(:status, :startTime)
+    DateTime.parse(str).to_time.to_i
+  end
+
+  def pod_status_from_json(json_data)
+    container_statuses = json_data.dig(:status, :containerStatuses)
+    json_state = container_statuses[0].dig(:state) # only support 1 container/pod
+
+    state = 'undetermined'
+    state = 'running' unless json_state.dig(:running).nil?
+    state = 'completed' unless json_state.dig(:terminated).nil?
+    state = 'queued' unless json_state.dig(:waiting).nil?
+
+    OodCore::Job::Status.new(state: state)
+  end
+
+end
